Start implementing joystick support.
This is obviously the most intuitive way to control a robot, so we
should support it. An XBox controller is readily available and plugs
straight into a USB port. It's also supported by pygame.

This just sends commands when it detects the stick is off-centre. There
is no analogue control and the 'stop' command is never sent. The whole
control structure really needs adjusting for the idea that we have
several potentially-conflicting inputs.